+++ packages/core/contracts/financial-templates/optimistic-distributor/OptimisticDistributor.sol
@@ -11,7 +11,6 @@ import "../../oracle/implementation/Constants.sol";
 import "../../oracle/interfaces/FinderInterface.sol";
 import "../../oracle/interfaces/IdentifierWhitelistInterface.sol";
 import "../../oracle/interfaces/OptimisticOracleInterface.sol";
-import "../../oracle/interfaces/StoreInterface.sol";
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
@@ -73,13 +72,17 @@ contract OptimisticDistributor is Lockable, MultiCaller, Testable {
     // Rewards are stored in dynamic array.
     Reward[] public rewards;
 
+    // Immutable variables used to validate input parameters when funding new rewards.
+    uint256 public immutable maximumFundingPeriod;
+    uint256 public immutable maximumProposerBond;
+
     // Proposals are mapped to hash of their identifier, timestamp and ancillaryData, so that they can be addressed
     // from OptimisticOracle callback function.
     mapping(bytes32 => Proposal) public proposals;
 
     // Immutable variables provided at deployment.
     FinderInterface public immutable finder;
-    IERC20 public bondToken; // This cannot be declared immutable as bondToken needs to be checked against whitelist.
+    IERC20 public immutable bondToken;
 
     // Merkle Distributor is automatically deployed on constructor and owned by this contract.
     MerkleDistributor public immutable merkleDistributor;
@@ -126,19 +129,25 @@ contract OptimisticDistributor is Lockable, MultiCaller, Testable {
 
     /**
      * @notice Constructor.
-     * @param _bondToken ERC20 token that the bond is paid in.
      * @param _finder Finder to look up UMA contract addresses.
+     * @param _bondToken ERC20 token that the bond is paid in.
      * @param _timer Contract that stores the current time in a testing environment.
+     * @param _maximumFundingPeriod Maximum period for reward funding (proposals allowed only afterwards).
+     * @param _maximumProposerBond Maximum allowed Optimistic Oracle proposer bond amount.
      */
     constructor(
         FinderInterface _finder,
         IERC20 _bondToken,
-        address _timer
+        address _timer,
+        uint256 _maximumFundingPeriod,
+        uint256 _maximumProposerBond
     ) Testable(_timer) {
         finder = _finder;
         require(_getCollateralWhitelist().isOnWhitelist(address(_bondToken)), "Bond token not supported");
         bondToken = _bondToken;
         syncUmaEcosystemParams();
+        maximumFundingPeriod = _maximumFundingPeriod;
+        maximumProposerBond = _maximumProposerBond;
         merkleDistributor = new MerkleDistributor();
     }
 
@@ -169,6 +178,8 @@ contract OptimisticDistributor is Lockable, MultiCaller, Testable {
         IERC20 rewardToken,
         bytes calldata customAncillaryData
     ) external nonReentrant() {
+        require(earliestProposalTimestamp <= getCurrentTime() + maximumFundingPeriod, "Too long till proposal opening");
+        require(optimisticOracleProposerBond <= maximumProposerBond, "OO proposer bond too high");
         require(_getIdentifierWhitelist().isIdentifierSupported(priceIdentifier), "Identifier not registered");
         require(_ancillaryDataWithinLimits(customAncillaryData), "Ancillary data too long");
         require(optimisticOracleLivenessTime >= MINIMUM_LIVENESS, "OO liveness too small");
@@ -361,10 +372,10 @@ contract OptimisticDistributor is Lockable, MultiCaller, Testable {
      ********************************************/
 
     /**
-     * @notice Updates the address stored in this contract for the OptimisticOracle and the Store to the latest
-     * versions set in the Finder. Also pull finalFee from Store contract.
-     * @dev There is no risk of leaving this function public for anyone to call as in all cases we want the addresses
-     * in this contract to map to the latest version in the Finder and store the latest final fee.
+     * @notice Updates the address stored in this contract for the OptimisticOracle to the latest version set
+     * in the Finder.
+     * @dev There is no risk of leaving this function public for anyone to call as in all cases we want the address of
+     * OptimisticOracle in this contract to map to the latest version in the Finder.
      */
     function syncUmaEcosystemParams() public nonReentrant() {
         optimisticOracle = _getOptimisticOracle();
@@ -402,10 +413,6 @@ contract OptimisticDistributor is Lockable, MultiCaller, Testable {
     /********************************************
      *            INTERNAL FUNCTIONS            *
      ********************************************/
-
-    function _getStore() internal view returns (StoreInterface) {
-        return StoreInterface(finder.getImplementationAddress(OracleInterfaces.Store));
-    }
 
     function _getOptimisticOracle() internal view returns (OptimisticOracleInterface) {
         return OptimisticOracleInterface(finder.getImplementationAddress(OracleInterfaces.OptimisticOracle));