+++ packages/contracts-node/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.7](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.3.6...@uma/contracts-node@0.3.7) (2022-05-27)
+
+**Note:** Version bump only for package @uma/contracts-node
+
 ## [0.3.6](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.3.5...@uma/contracts-node@0.3.6) (2022-05-13)
 
 **Note:** Version bump only for package @uma/contracts-node